filter option for staff
